chore: move all unit tests to __tests__ dirs and update paths
