+++ src/__tests__/hello.spec.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react'
 import React from 'react'
 import { describe, it, expect } from 'vitest'
-import Hello from './hello'
+import Hello from '../hello'
 
 describe('Hello', () => {
     it('displays the name', () => {
